+++ .travis/install_python.sh
@@ -221,6 +221,8 @@ function install_mac_pypy3 {
 
 if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
     install_python
-    virtualenv venv -p $PYTHON_EXE
-    source venv/bin/activate
+    if [[ -n "$PYTHON_EXE" ]]; then
+        virtualenv venv -p $PYTHON_EXE
+        source venv/bin/activate
+    fi
 fi