+++ tcod/noise.py
@@ -1,8 +1,11 @@
 
-from __future__ import absolute_import as _
+from __future__ import absolute_import
+
+import operator
+
+import numpy as np
 
 from tcod.tcod import _cdata
-from tcod.tcod import _CDataWrapper
 from tcod.libtcod import ffi, lib
 
 # Noise implementation constants
@@ -10,7 +13,7 @@ SIMPLE = 0
 FBM = 1
 TURBULENCE = 2
 
-class Noise(_CDataWrapper):
+class Noise(object):
     """
     .. versionadded:: 2.0
 
@@ -23,7 +26,7 @@ class Noise(_CDataWrapper):
     Not used with NOISE_IMP_SIMPLE.
 
     Args:
-        dimentions (int): Must be from 1 to 4.
+        dimensions (int): Must be from 1 to 4.
         algorithm (int): Defaults to NOISE_SIMPLEX
         implementation (int): Defaults to NOISE_IMP_SIMPLE
         hurst (float): The hurst exponent.  Should be in the 0.0-1.0 range.
@@ -32,31 +35,29 @@ class Noise(_CDataWrapper):
                          implementations.
         rand (Optional[Random]): A Random instance, or None.
     """
-    def __init__(self, *args, **kargs):
-        self.octaves = 4
-        self.implementation = SIMPLE
-        self._cdata_random = None # keep alive the random cdata instance
-        self._algorithm = None
-        self._dimentions = None
-        self._hurst = None
-        self._lacunarity = None
-        super(Noise, self).__init__(*args, **kargs)
-        if not self.cdata:
-            self._init(*args, **kargs)
 
-    def _init(self, dimentions, algorithm=2, implementation=SIMPLE,
-              hurst=0.5, lacunarity=2.0, octaves=4, rand=None):
-        self._cdata_random = _cdata(rand)
-        self.implementation = implementation
-        self._dimentions = dimentions
+    def __init__(self, dimensions, algorithm=2, implementation=SIMPLE,
+                 hurst=0.5, lacunarity=2.0, octaves=4, rand=None):
+        if not 0 < dimensions <= 4:
+            raise ValueError('dimensions must be in range 0 < n <= 4, got %r' %
+                             (dimensions,))
+        self._random = rand
+        self._random_c = _cdata(rand)
+        self._algorithm = algorithm
         self._hurst = hurst
         self._lacunarity = lacunarity
-        self.octaves = octaves
-        self.cdata = ffi.gc(lib.TCOD_noise_new(self._dimentions, self._hurst,
-                                               self._lacunarity,
-                                               self._cdata_random),
-                            lib.TCOD_noise_delete)
-        self.algorithm = algorithm
+        self.noise_c = ffi.gc(
+            lib.TCOD_noise_new(dimensions, hurst, lacunarity, self._random_c),
+            lib.TCOD_noise_delete)
+        self._tdl_noise_c = ffi.new('TDLNoise*', (self.noise_c,
+                                                  dimensions,
+                                                  0,
+                                                  octaves))
+        self.implementation = implementation # sanity check
+
+    @property
+    def dimensions(self):
+        return self._tdl_noise_c.dimensions
 
     @property
     def algorithm(self):
@@ -64,11 +65,16 @@ class Noise(_CDataWrapper):
     @algorithm.setter
     def algorithm(self, value):
         self._algorithm = value
-        lib.TCOD_noise_set_type(self.cdata, value)
+        lib.TCOD_noise_set_type(self.noise_c, value)
 
     @property
-    def dimentions(self):
-        return self._dimentions
+    def implementation(self):
+        return self._tdl_noise_c.implementation
+    @implementation.setter
+    def implementation(self, value):
+        if not 0 <= value < 3:
+            raise ValueError('%r is not a valid implementation. ' % (value,))
+        self._tdl_noise_c.implementation = value
 
     @property
     def hurst(self):
@@ -77,6 +83,13 @@ class Noise(_CDataWrapper):
     @property
     def lacunarity(self):
         return self._lacunarity
+
+    @property
+    def octaves(self):
+        return self._tdl_noise_c.octaves
+    @octaves.setter
+    def octaves(self, value):
+        self._tdl_noise_c.octaves = value
 
     def get_point(self, x=0, y=0, z=0, w=0):
         """Return the noise value at the (x, y, z, w) point.
@@ -87,13 +100,55 @@ class Noise(_CDataWrapper):
             z (float): The position on the 3rd axis.
             w (float): The position on the 4th axis.
         """
-        if self.implementation == SIMPLE:
-            return lib.TCOD_noise_get(self.cdata, (x, y, z, w))
-        elif self.implementation == FBM:
-            return lib.TCOD_noise_get_fbm(self.cdata, (x, y, z, w),
-                                          self.octaves)
-        elif self.implementation == TURBULENCE:
-            return lib.TCOD_noise_get_turbulence(self.cdata, (x, y, z, w),
-                                                 self.octaves)
-        raise RuntimeError('implementation must be one of tcod.NOISE_IMP_*')
+        return lib.NoiseGetSample(self._tdl_noise_c, (x, y, z, w))
 
+    def sample_mgrid(self, mgrid):
+        """Sample a mesh-grid array and return the result.
+
+        Args:
+            mgrid (numpy.ndarray): A mesh-grid array of points to sample.
+
+        Returns:
+            numpy.ndarray: A float32 array of sampled points
+                           with the shape: ``mgrid.shape[:-1]``.
+
+        .. versionadded:: 2.2
+        """
+        mgrid = np.ascontiguousarray(mgrid, np.float32)
+        if mgrid.shape[0] != self.dimensions:
+            raise ValueError('mgrid.shape[0] must equal self.dimensions, '
+                             '%r[0] != %r' % (mgrid.shape, self.dimensions))
+        out = np.ndarray(mgrid.shape[1:], np.float32)
+        if mgrid.shape[1:] != out.shape:
+            raise ValueError('mgrid.shape[1:] must equal out.shape, '
+                             '%r[1:] != %r' % (mgrid.shape, out.shape))
+        lib.NoiseSampleMeshGrid(self._tdl_noise_c, out.size,
+                                ffi.cast('float*', mgrid.ctypes.data),
+                                ffi.cast('float*', out.ctypes.data))
+        return out
+
+    def sample_ogrid(self, ogrid):
+        """Sample an open mesh-grid array and return the result.
+
+        Args
+            ogrid (Sequence[numpy.ndarray]): An open mesh-grid.
+
+        Returns:
+            numpy.ndarray:  A float32 array of sampled points.  Shape is based
+                            on the lengths of the open mesh-grid arrays.
+
+        .. versionadded:: 2.2
+        """
+        if len(ogrid) != self.dimensions:
+            raise ValueError('len(ogrid) must equal self.dimensions, '
+                             '%r != %r' % (len(ogrid), self.dimensions))
+        ogrids = [np.ascontiguousarray(array, np.float32) for array in ogrid]
+        out = np.ndarray([array.size for array in ogrids], np.float32)
+        lib.NoiseSampleOpenMeshGrid(
+            self._tdl_noise_c,
+            len(ogrids),
+            out.shape,
+            [ffi.cast('float*', array.ctypes.data) for array in ogrids],
+            ffi.cast('float*', out.ctypes.data),
+            )
+        return out